requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
waitress==2.0.0
mysql-connector-python==8.0.19
